Updated sequencer to use size_t rather than uint32

+++ core/hpp/sequencer.hpp
@@ -1,17 +1,16 @@
 #ifndef SEQUENCER_HPP
 #define SEQUENCER_HPP
 
-#include <cstdint>
 #include "trail.hpp"
 #include "tracked.hpp"
 
 struct sequencer {
     sequencer(trail&);
-    uint32_t operator()();
+    size_t operator()();
 #ifndef DEBUG
 private:
 #endif
-    tracked<uint32_t> index;
+    tracked<size_t> index;
 };
 
 #endif
